PyStrategy is now a full-fledged python package

+++ tests/market_data_simulation_manager.hpp
@@ -12,14 +12,14 @@ using BufferPtr = char*;
 TEST(MDRow, AllOperations)
 {
     std::vector<MDCustomUpdate> updates1(5, MDCustomUpdate());
-    for (int i = 0; i < updates1.size(); ++i)
+    for (size_t i = 0; i < updates1.size(); ++i)
     {
         updates1[i].EventTimestamp = i*2 + 1;
         updates1[i].Payload = i;
     }
 
     MDRow row1(updates1);
-    for(int i = 0; i < row1.size(); ++i)
+    for(size_t i = 0; i < row1.size(); ++i)
     {
         auto r = row1[i];
         EXPECT_EQ(r->EventTimestamp, i*2 + 1);
@@ -32,9 +32,9 @@ TEST(MarketDataSimulationManagerTests, IteratorInc)
 {
     std::vector<MDCustomUpdate> updates1(5, MDCustomUpdate());
     std::vector<MDCustomUpdate> updates2(5, MDCustomUpdate());
-    for (int i = 0; i < updates1.size(); ++i)
+    for (size_t i = 0; i < updates1.size(); ++i)
         updates1[i].EventTimestamp = i*2 + 1;
-    for (int i = 0; i < updates2.size(); ++i)
+    for (size_t i = 0; i < updates2.size(); ++i)
         updates2[i].EventTimestamp = i*2;
     MarketDataSimulationManager manager(std::vector<MDRow>{updates1, updates2});
     auto iter = manager.begin();
@@ -53,11 +53,11 @@ TEST(MarketDataSimulationManagerTests, IteratorPlus)
     std::vector<MDCustomUpdate> updates1(5, MDCustomUpdate());
     std::vector<MDCustomUpdate> updates2(5, MDCustomUpdate());
     std::vector<MDTrade> updates3(5, MDTrade());
-    for (int i = 0; i < updates1.size(); ++i)
+    for (size_t i = 0; i < updates1.size(); ++i)
         updates1[i].EventTimestamp = i*3;
-    for (int i = 0; i < updates2.size(); ++i)
+    for (size_t i = 0; i < updates2.size(); ++i)
         updates2[i].EventTimestamp = i*3 + 1;
-    for (int i = 0; i < updates3.size(); ++i)
+    for (size_t i = 0; i < updates3.size(); ++i)
     {
         updates3[i].EventTimestamp = i*3 + 2;
         updates3[i].Price = 2;
